Extract CLI to module
