+++ src/cli.rs
@@ -1,0 +1,22 @@
+use clap::Parser;
+
+/// Measure performance of a program across git commits.
+#[derive(Parser, Debug)]
+#[command(version, about, long_about = None)]
+pub struct Args {
+    /// Program to run
+    #[arg(short, long)]
+    pub program: String,
+
+    /// Arguments to pass to program
+    #[arg(short, long)]
+    pub arg: Option<Vec<String>>,
+
+    /// Working directory for program execution
+    #[arg(short, long)]
+    pub working_dir: Option<String>,
+
+    /// Whether to show program output
+    #[arg(long, action)]
+    pub show_output: bool,
+}
